Add lib & bin to allow shuttle next to other means
